requirements.txt
# Kafka
kafka-python==2.0.2

# WebSocket
websocket-client==1.7.0

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Data Processing
pandas==2.2.0
numpy==1.26.3

# Visualization
streamlit==1.31.0
plotly==5.18.0

# Configuration
python-dotenv==1.0.1

# Utilities
requests==2.31.0

# Testing
pytest==8.0.0
pytest-cov==4.1.0
